# docs/swagger.yaml
definitions:
  internal.CompleteStep:
    properties:
      stepid:
        description: TODO вообще правильно ИД пользователя не передавать, а брать
          из авторизации, но тогда будет сложно тестировать, с учетом того, что это
          тестовое задание, то будем передавать
        type: integer
      userid:
        description: Идентификатор пользователя выполневшего шаг
        type: integer
    type: object
  internal.DeleteUserStruct:
    properties:
      id:
        type: integer
    type: object
  internal.NewCompleteSteps:
    description: NewCompleteSteps  json для отметки о выполнении шага задания пользователем
    properties:
      CompleteSteps:
        description: Идентификатор задания
        items:
          $ref: '#/definitions/internal.CompleteStep'
        type: array
    type: object
  internal.NewQuest:
    description: NewQuest json для создания задания с шагами
    properties:
      Name:
        description: Имя задания
        type: string
      QuestSteps:
        description: Шаги задания
        items:
          $ref: '#/definitions/internal.NewQuestStep'
        type: array
      id:
        description: Идентификатор задания
        type: integer
    type: object
  internal.NewQuestStep:
    properties:
      Bonus:
        description: Бонус за задание
        type: integer
      IsMulti:
        description: Признак того, что шаг можно выполнять несколько раз
        type: boolean
      QuestId:
        description: Идентификатор задания. При создании методом CreateQuest, значение
          будет проигнорировано, т.к. будет подставляться идентификатор создаваемого
          задания
        type: integer
      StepName:
        description: Описание шага
        type: string
      id:
        description: Идентификатор задания
        type: integer
    type: object
  internal.NewQuestSteps:
    description: NewQuestStep json для создания шага задания
    properties:
      QuestSteps:
        description: Идентификатор задания
        items:
          $ref: '#/definitions/internal.NewQuestStep'
        type: array
    type: object
  internal.Quests:
    description: Quests json информация о заданиях и их шагов
    properties:
      Id:
        description: ИД задания
        type: string
      QuestName:
        description: Имя выполненного задания пользователем
        type: string
      Steps:
        description: Шаги задания
        items:
          $ref: '#/definitions/internal.Steps'
        type: array
    type: object
  internal.Steps:
    properties:
      Bonus:
        description: Бонус за выполнение шага
        type: integer
      Id:
        description: ИД шага
        type: integer
      StepName:
        description: Имя шага
        type: string
      isMulti:
        description: Признак того, что шаг можно выполнять повторно
        type: boolean
    type: object
  internal.UpdateQuestStep:
    properties:
      Bonus:
        description: Бонус за задание
        type: integer
      IsMulti:
        description: Признак того, что шаг можно выполнять несколько раз
        type: boolean
      id:
        description: Идентификатор задания
        type: integer
    type: object
  internal.UpdateQuestSteps:
    description: UpdateQuestSteps json для обновления шагов заданий
    properties:
      QuestSteps:
        description: Идентификатор задания
        items:
          $ref: '#/definitions/internal.UpdateQuestStep'
        type: array
    type: object
  internal.User:
    properties:
      id:
        description: идентификатор пользователя
        type: integer
      password:
        description: пароль пользователя
        type: string
      userIsAdmin:
        description: признак того, что пользователь является администратором
        type: boolean
      username:
        description: имя пользователя
        type: string
    type: object
  internal.UserBonus:
    description: UserBonus json для получения история выполнения заданий и их шагов
    properties:
      ComplitedQuests:
        description: Список заданий в которых участвовал пользователь
        items:
          $ref: '#/definitions/internal.UserCompletedQuest'
        type: array
      TotalBonus:
        description: Общий бонусный счет пользователя
        type: integer
    type: object
  internal.UserCompletedQuest:
    properties:
      AllStepsCount:
        description: Кол-во шагов, доступное в задании
        type: integer
      Bonus:
        description: Сумма Бонусов за выполненные задания
        type: integer
      CompletedSteps:
        description: Выполненные шаги пользователем
        items:
          $ref: '#/definitions/internal.UserCompletedSteps'
        type: array
      CompletedStepsCount:
        description: Кол-во выполненных шагов заданий пользователем
        type: integer
      QuestId:
        description: ИД задания
        type: string
      QuestName:
        description: Имя выполненного задания пользователем
        type: string
    type: object
  internal.UserCompletedSteps:
    properties:
      Count:
        description: Кол-во выполнений шага
        type: integer
      StepName:
        description: Имя выполненного шага
        type: string
      UserBonusStep:
        description: Бонус пользователя за выполнение шага
        type: integer
    type: object
host: localhost:8081
info:
  contact: {}
  description: API Server for Quests Application
  title: quests API
  version: "1.0"
paths:
  /History/CompleteSteps:
    post:
      consumes:
      - application/json
      description: Устанавливает признак выполнения шага для пользователя
      operationId: CompleteSteps
      parameters:
      - description: обновленная информация о шагах задания
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal.NewCompleteSteps'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.NewCompleteSteps'
      security:
      - BasicAuth: []
      summary: установить пометки у выполнении для шагов
      tags:
      - history
  /History/GetHistory:
    get:
      consumes:
      - application/json
      description: Показывает историю выполнения заданий для пользователя
      operationId: GetHistory
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.UserBonus'
      security:
      - BasicAuth: []
      summary: История по пользователю
      tags:
      - history
  /Quests/CreateQuest:
    post:
      consumes:
      - application/json
      description: Создает новое задание
      operationId: CreateQuest
      parameters:
      - description: информация о задании
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal.NewQuest'
      responses:
        "200":
          description: OK
      security:
      - BasicAuth: []
      summary: Добавить задание
      tags:
      - quests
  /Quests/CreateQuestSteps:
    post:
      consumes:
      - application/json
      description: Добавляет информацию о новых шагах к заданию
      operationId: CreateQuestSteps
      parameters:
      - description: информация о шагах задания
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal.NewQuestSteps'
      responses:
        "200":
          description: OK
      security:
      - BasicAuth: []
      summary: Добавляет информацию о новых шагах к заданию
      tags:
      - quests
  /Quests/GetQuests:
    get:
      consumes:
      - application/json
      description: Показывает все задания в системе
      operationId: GetQuests
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal.Quests'
            type: array
      security:
      - BasicAuth: []
      summary: Показывает все задания
      tags:
      - quests
  /Quests/UpdateQuestSteps:
    post:
      consumes:
      - application/json
      description: Обновляет информацию о шагах заданий
      operationId: UpdateQuestSteps
      parameters:
      - description: обновленная информация о шагах задания
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal.UpdateQuestSteps'
      responses:
        "200":
          description: OK
      security:
      - BasicAuth: []
      tags:
      - quests
  /Users/CreateUser:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя приложения
      operationId: CreateUser
      parameters:
      - description: Информация о пользователе
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal.User'
      responses: {}
      security:
      - BasicAuth: []
      summary: Создать пользователя
      tags:
      - user
  /Users/DeleteUser:
    delete:
      consumes:
      - application/json
      description: Удаляет пользователя приложения
      operationId: DeleteUser
      parameters:
      - description: Идентификатор пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal.DeleteUserStruct'
      responses: {}
      security:
      - BasicAuth: []
      summary: Удалить пользователя
      tags:
      - user
  /Users/GetAllUser:
    get:
      consumes:
      - application/json
      description: Показать всех пользователей
      operationId: GetAllUser
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal.User'
            type: array
      security:
      - BasicAuth: []
      summary: Показать всех пользователей
      tags:
      - user
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"